feat(utxorpc): add v1beta spec support alongside v1alpha
Bumps utxorpc-spec to 0.19 and exposes both `pallas_utxorpc::v1alpha`
and `pallas_utxorpc::v1beta` mappers. The crate root keeps re-exporting
v1alpha items (Mapper, spec) under the default `v1alpha` feature, so
existing call sites compile unchanged. v1beta is gated behind an
opt-in `v1beta` feature; both features are additive.

Shared mapping logic (~80% of the surface — all certificates, pparams,
plutus data, witnesses, scripts, blocks) lives in a single
`impl_cardano_mapper_shared!` macro instantiated once per version. Methods
that diverge between specs (map_native_script, map_tx_datum, map_tx_output,
map_asset, map_policy_assets, map_conway_gov_action, map_tx) are defined
per-version. v1beta-only mappers (BootstrapWitness, Vote, VotingProcedure,
VoterVotes) live next to the rest of the v1beta module.

Adds cross-version wire-compatibility tests (PParams, Certificate,
PoolRegistrationCert, Metadata, PlutusData) that act as a tripwire if
upstream diverges any shared message in a future release.
