+++ pallas-utxorpc/Cargo.toml
@@ -12,7 +12,7 @@ authors = ["Santiago Carmuega <santiago@carmuega.me>"]
 
 [dependencies]
 # utxorpc-spec = { path = "../../../utxorpc/spec/gen/rust" }
-utxorpc-spec = { version = "0.18.1" }
+utxorpc-spec = { version = "0.19", default-features = false }
 
 pallas-traverse = { version = "=1.0.0-alpha.6", path = "../pallas-traverse" }
 pallas-primitives = { version = "=1.0.0-alpha.6", path = "../pallas-primitives" }
@@ -21,7 +21,13 @@ pallas-crypto = { version = "=1.0.0-alpha.6", path = "../pallas-crypto" }
 prost-types = "0.13.1"
 pallas-validate = { version = "=1.0.0-alpha.6", path = "../pallas-validate" }
 
+[features]
+default = ["v1alpha"]
+v1alpha = ["utxorpc-spec/utxorpc-v1alpha-cardano"]
+v1beta = ["utxorpc-spec/utxorpc-v1beta-cardano"]
+
 [dev-dependencies]
 hex = "0.4.3"
 serde_json = "1.0.120"
 pretty_assertions = "1.4.0"
+prost = "0.13"
